requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.30.0
pillow>=10.0.0
numpy>=1.24.0
opencv-python-headless>=4.8.0
torch
torchvision
requests>=2.31.0
python-multipart>=0.0.9
tqdm>=4.66.0
